Move bullet donation to archon
