+++ logic/ArchonLogic.java
@@ -16,6 +16,9 @@ import battlecode.common.TreeInfo;
 
 public class ArchonLogic extends RobotLogic {
 
+	private static final float DONATE_AFTER = 500; // bullets
+	private static final float DONATE_PERCENTAGE = 0.10f;
+	
 	// Prioritized build directions
 	private List<Direction> buildDirs = new ArrayList<>(
 			Arrays.asList(new Direction[] { Direction.getNorth(), Navigation.NORTH_EAST, Navigation.NORTH_WEST,
@@ -37,6 +40,8 @@ public class ArchonLogic extends RobotLogic {
 	@Override
 	public void logic() {
 		
+		donateToWin();
+		
 		if (rc.getLocation().distanceTo(enemySpawn) < 20){
 			nav.moveBest(enemySpawn.directionTo(rc.getLocation()));
 		}
@@ -45,6 +50,29 @@ public class ArchonLogic extends RobotLogic {
 		nav.dodgeBullets();
 	}
 
+	void donateToWin() {
+		try {
+			if (rc.getRoundNum() == rc.getRoundLimit() - 1) {
+
+				// End game. Just donate all.
+				rc.donate(rc.getTeamBullets());
+
+			} else {
+				// If we can win... win.
+				if (rc.getTeamVictoryPoints()
+						+ (int) (rc.getTeamBullets() / 10) >= GameConstants.VICTORY_POINTS_TO_WIN) {
+
+					rc.donate(rc.getTeamBullets());
+
+				} else if (rc.getTeamBullets() > DONATE_AFTER) {
+					rc.donate(rc.getTeamBullets() * DONATE_PERCENTAGE);
+				}
+			}
+		} catch (GameActionException e) {
+			e.printStackTrace();
+		}
+	}
+	
 	void buildBase() {
 		Direction built = null;
 		for (Direction dir : buildDirs) {
